third_party/blink/web_tests/platform/mac/inspector-protocol/css/css-resolve-values-expected.txt

Test css.resolveValue method

Running test: testNotExistentNode
Test non existent node
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"code":-32000,"message":"Could not find node with given id"}

Running test: testInvalidProperty
Test invalid property name
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"code":-32000,"message":"Could not resolve property name."}

Running test: testResolveValuesSimple
Test resolveValues for width property
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"results":["","5px","30px","6px","6px"]}

Running test: testRelativeUnits
Relative length units to absolute test
{"input": [1em,1ex,1cap,1ch,1ic,1lh,1rcap,1rch,1rem,1rex,1ric,1rlh,1vw,1vh ,1vi,1vb,1vmin,1vmax]}
{"results":["5px","2.24365px","3.30811px","2.5px","5px","6px","19.8486px","15px","30px","13.4619px","30px","36px","8px","6px","8px","6px","6px","8px"]}

Running test: testRelativeUnitsOuter
Relative length units to absolute for parent element test
{"input": [1em,1ex,1cap,1ch,1ic,1lh,1rcap,1rch,1rem,1rex,1ric,1rlh,1vw,1vh ,1vi,1vb,1vmin,1vmax]}
{"results":["10px","4.4873px","6.61621px","5px","10px","13px","19.8486px","15px","30px","13.4619px","30px","36px","8px","6px","8px","6px","6px","8px"]}

Running test: testRelativeUnitsFontSize
Relative length units to absolute font-size property test
{"input": [1em,1ex,1cap,1ch,1ic,1lh,1rcap,1rch,1rem,1rex,1ric,1rlh,1vw,1vh ,1vi,1vb,1vmin,1vmax]}
{"results":["10px","4.4873px","6.61621px","5px","10px","13px","19.8486px","15px","30px","13.4619px","30px","36px","8px","6px","8px","6px","6px","8px"]}

Running test: testExpressionsEvaluation
Evaluate expression test
{"input": [calc(1em + 10px),calc(1em + 3em),calc(3px + 2.54cm),clamp(10px, calc(10rem + 10rex), 30px),max(100px, 30em),calc(100px * cos(60deg))]}
{"results":["15px","20px","99px","30px","150px","50px"]}

Running test: testCSSWideKeywords
Test CSS-wide keywords
{"input": [initial,inherit,unset]}
{"results":["16px","10px","10px"]}

Running test: testColorVaslues
Test <color> values
{"input": [aqua,peachpuff,blanchedalmond,rgb(255, 0, 0),#0f5ffe,color-mix(in srgb, plum, #f00)]}
{"results":["rgb(0, 255, 255)","rgb(255, 218, 185)","rgb(255, 235, 205)","rgb(255, 0, 0)","rgb(15, 95, 254)","color(srgb 0.933333 0.313726 0.433333)"]}

Running test: testOnlyInvalidValues
Invalid values test
{"input": [calc(,em,calc(10 + 20),red,calc(10ms + 5s)]}
{"results":["","","","",""]}

Running test: testCustomProperty
Test resolveValues on custom property
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"results":["invalid","1em","1rem","calc(3px + 3px)","calc(1em + 1px)"]}

Running test: testRegisterCustomProperty
Test resolveValues on register custom property
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"results":["","10px","30px","6px","11px"]}

Running test: testPseudoElement
Pseudo element test
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"results":["","3px","30px","6px","4px"]}

Running test: testNonExistentPseudoElement
Non existent pseudo element test
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"code":-32000,"message":"Could not retrieve pseudo element."}

Running test: testPseudoElementsNoElement
Pseudo element with no content
{"input": [invalid,1em,1rem,calc(3px + 3px),calc(1em + 1px)]}
{"code":-32000,"message":"Could not retrieve pseudo element."}
